# eliasfano32: optimise reverse iterator