--- a/erigon-lib/recsplit/eliasfano32/elias_fano_fuzz_test.go
+++ b/erigon-lib/recsplit/eliasfano32/elias_fano_fuzz_test.go
@@ -74,7 +74,6 @@ func FuzzSingleEliasFano(f *testing.F) {
 
 		i--
 		rit := ef.ReverseIterator()
-		//rit.real = keys
 		for rit.HasNext() {
 			v, err := rit.Next()
 			if err != nil {